--- a/demo/blocks_essay/run.py
+++ b/demo/blocks_essay/run.py
@@ -20,4 +20,4 @@ with gr.Blocks() as demo:
 
 
 if __name__ == "__main__":
-    demo.launch()
+    demo.launch(prevent_thread_lock=True)