--- a/gradio/context.py
+++ b/gradio/context.py
@@ -12,3 +12,4 @@ class Context:
     root_block: Blocks | None = None  # The current root block that holds all blocks.
     block: BlockContext | None = None  # The current block that children are added to.
     id: int = 0  # Running id to uniquely refer to any block that gets defined
+    ip_address: str | None = None