--- a/test/test_utils.py
+++ b/test/test_utils.py
@@ -12,6 +12,7 @@ from httpx import AsyncClient, Response
 from pydantic import BaseModel
 from typing_extensions import Literal
 
+from gradio.context import Context
 from gradio.test_data.blocks_configs import (
     XRAY_CONFIG,
     XRAY_CONFIG_DIFF_IDS,
@@ -65,12 +66,12 @@ class TestUtils:
     @mock.patch("requests.post")
     def test_error_analytics_doesnt_crash_on_connection_error(self, mock_post):
         mock_post.side_effect = requests.ConnectionError()
-        error_analytics("placeholder", "placeholder")
+        error_analytics("placeholder")
         mock_post.assert_called()
 
     @mock.patch("requests.post")
     def test_error_analytics_successful(self, mock_post):
-        error_analytics("placeholder", "placeholder")
+        error_analytics("placeholder")
         mock_post.assert_called()
 
     @mock.patch("requests.post")
@@ -106,6 +107,7 @@ class TestUtils:
 class TestIPAddress:
     @pytest.mark.flaky
     def test_get_ip(self):
+        Context.ip_address = None
         ip = get_local_ip_address()
         if ip == "No internet connection":
             return
@@ -113,6 +115,7 @@ class TestIPAddress:
 
     @mock.patch("requests.get")
     def test_get_ip_without_internet(self, mock_get):
+        Context.ip_address = None
         mock_get.side_effect = requests.ConnectionError()
         ip = get_local_ip_address()
         assert ip == "No internet connection"