@abidlabs 2023-01-25
I'm curious what the performance difference is on faster wifi. You can run this command to profile a demo:

```bash
python -m cProfile  -s time -o profile.txt .\demo\blocks_essay\run.py
```