@abidlabs 2023-01-25
Fixed tests and confirmed that analytics still show up on Segment: 

<img width="409" alt="image" src="https://user-images.githubusercontent.com/1778297/214563660-a359c114-3a55-424a-8e21-ff6d26a3541a.png">