@FarukOzderim 2023-01-25
Hello, @abidlabs @freddyaboulton!

WDYT about adding these background requests [Startup-events].(https://github.com/gradio-app/gradio/blob/7fc3e750b04290c2066758f8697c575cef3f6441/gradio/routes.py#L475) 

1. It would be more tidy and structured.
2. They could run as async requests in the background, having a lighter weight on startup.