@abidlabs 2023-01-26
But some of these (e.g. the initiated analytics) happen before the Interface/Blocks is launched (they happen when the Interface object is created) so I don't think this would work @FarukOzderim?