@FarukOzderim 2023-01-26
Hmm, to be sure, they are not needed immediately right?
Then all of the prior calls could be stored inside a list which will be called in startup.