@abidlabs 2023-01-28
Hmm but what if the app is never launched? The purpose of the _initiated analytics_ was to log an event when the `Interface` object is created so that we could see if there are certain interfaces that users are not launching (e.g. using as functions, etc.)